add_predictions {modelr}	R Documentation

Add predictions to a data frame

Description

Add predictions to a data frame

Usage

add_predictions(data, model, var = "pred")

spread_predictions(data, ...)

gather_predictions(data, ..., .pred = "pred", .model = "model")

Arguments

data
A data frame used to generate the predictions.

model, var
add_predictions takes a single model; the output column will be called pred

...
gather_predictions and spread_predictions take multiple models. The name will be taken from either the argument name of the name of the model.

.pred, .model
The variable names used by gather_predictions.

Value

A data frame. add_prediction adds a single new column, .pred, to the input data. spread_predictions adds one column for each model. gather_predictions adds two columns .model and .pred, and repeats the input rows for each model.

Examples

df <- tibble::data_frame(
  x = sort(runif(100)),
  y = 5 * x + 0.5 * x ^ 2 + 3 + rnorm(length(x))
)
plot(df)

m1 <- lm(y ~ x, data = df)
grid <- data.frame(x = seq(0, 1, length = 10))
grid %>% add_predictions(m1)

m2 <- lm(y ~ poly(x, 2), data = df)
grid %>% spread_predictions(m1, m2)
grid %>% gather_predictions(m1, m2)

[Package modelr version 0.1.2 Index]