heights {modelr}	R Documentation

Height and income data.

Description

You might have heard that taller people earn more. Is it true? You can try and answer the question by exploring this dataset extracted from the National Longitudinal Study, which is sponsored by the U.S. Bureau of Labor Statistics.

Usage

heights

Format

income

Yearly income. The top two percent of values were averaged and that average was used to replace all values in the top range.

height

Height, in feet

weight

Weight, in pounds

age

Age, in years, between 47 and 56.

marital

Marital status

sex

Sex

education

Years of education

afqt

Percentile score on Armed Forces Qualification Test.

Details

This contains data as at 2012.

[Package modelr version 0.1.2 Index]