modelr-package {modelr}	R Documentation

modelr: Modelling Functions that Work with the Pipe

Description

Functions for modelling that help you seamlessly integrate modelling into a pipeline of data manipulation and visualisation.

Author(s)

Maintainer: Hadley Wickham hadley@rstudio.com

Other contributors:

RStudio [copyright holder]

See Also

Useful links:

https://modelr.tidyverse.org

https://github.com/tidyverse/modelr

Report bugs at https://github.com/tidyverse/modelr/issues

[Package modelr version 0.1.8 Index]