A "lazy" resample.

Often you will resample a dataset hundreds or thousands of times. Storing the complete resample each time would be very inefficient so this class instead stores a "pointer" to the original dataset, and a vector of row indexes. To turn this into a regular data frame, call as.data.frame, to extract the indices, use as.integer.

Usage

resample(data, idx)

Arguments

data

The data frame

idx

A vector of integer indexes indicating which rows have been selected. These values should lie between 1 and nrow(data) but they are not checked by this function in the interests of performance.

See also

Examples

resample(mtcars, 1:10)
#> <resample [10 x 11]> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10

b <- resample_bootstrap(mtcars)
b
#> <resample [32 x 11]> 32, 5, 8, 12, 5, 4, 4, 29, 4, 17, ...
as.integer(b)
#>  [1] 32  5  8 12  5  4  4 29  4 17 30 15 14 27  9 24 22 14 12 13  3 28 15
#> [24] 31 13 13 17 22 29 31  8 21
as.data.frame(b)
#>                       mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Volvo 142E           21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Hornet Sportabout    18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Merc 240D            24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 450SE           16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Hornet Sportabout.1  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Hornet 4 Drive       21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet 4 Drive.1     21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Ford Pantera L       15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Hornet 4 Drive.2     21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Chrysler Imperial    14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Ferrari Dino         19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Cadillac Fleetwood   10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Merc 450SLC          15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Porsche 914-2        26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Merc 230             22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Camaro Z28           13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Dodge Challenger     15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Merc 450SLC.1        15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Merc 450SE.1         16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL           17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Datsun 710           22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Lotus Europa         30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Cadillac Fleetwood.1 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Maserati Bora        15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Merc 450SL.1         17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SL.2         17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Chrysler Imperial.1  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Dodge Challenger.1   15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Ford Pantera L.1     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Maserati Bora.1      15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Merc 240D.1          24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Toyota Corona        21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1

# Many modelling functions will do the coercion for you, so you can
# use a resample object directly in the data argument
lm(mpg ~ wt, data = b)
#> 
#> Call:
#> lm(formula = mpg ~ wt, data = b)
#> 
#> Coefficients:
#> (Intercept)           wt  
#>      32.984       -4.172  
#>