A "lazy" resample.

Often you will resample a dataset hundreds or thousands of times. Storing the complete resample each time would be very inefficient so this class instead stores a "pointer" to the original dataset, and a vector of row indexes. To turn this into a regular data frame, call as.data.frame, to extract the indices, use as.integer.

Usage

resample(data, idx)

Arguments

data

The data frame

idx

A vector of integer indexes indicating which rows have been selected. These values should lie between 1 and nrow(data) but they are not checked by this function in the interests of performance.

See also

Examples

resample(mtcars, 1:10)
#> <resample [10 x 11]> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10

b <- resample_bootstrap(mtcars)
b
#> <resample [32 x 11]> 8, 25, 27, 16, 16, 11, 27, 3, 14, 8, ...
as.integer(b)
#>  [1]  8 25 27 16 16 11 27  3 14  8 21  2 10  1 28  1 17 26 11 25 21 25  8
#> [24] 26 25 32 32 25 20 16 18 32
as.data.frame(b)
#>                        mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Merc 240D             24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Pontiac Firebird      19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Porsche 914-2         26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lincoln Continental   10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Lincoln Continental.1 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Merc 280C             17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Porsche 914-2.1       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Datsun 710            22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Merc 450SLC           15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Merc 240D.1           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Toyota Corona         21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Mazda RX4 Wag         21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Merc 280              19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Mazda RX4             21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Lotus Europa          30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Mazda RX4.1           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Chrysler Imperial     14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat X1-9             27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Merc 280C.1           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Pontiac Firebird.1    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Toyota Corona.1       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Pontiac Firebird.2    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Merc 240D.2           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Fiat X1-9.1           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Pontiac Firebird.3    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Volvo 142E            21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Volvo 142E.1          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Pontiac Firebird.4    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Toyota Corolla        33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Lincoln Continental.2 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Fiat 128              32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Volvo 142E.2          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2

# Many modelling functions will do the coercion for you, so you can
# use a resample object directly in the data argument
lm(mpg ~ wt, data = b)
#> 
#> Call:
#> lm(formula = mpg ~ wt, data = b)
#> 
#> Coefficients:
#> (Intercept)           wt  
#>      36.154       -4.666  
#>