resample_partition {modelr}	R Documentation

Generate an exclusive partitioning of a data frame

Description

Generate an exclusive partitioning of a data frame

Usage

resample_partition(data, p)

Arguments

data
A data frame

p
A named numeric vector giving where the value is the probability that an observation will be assigned to that group.

Examples

ex <- resample_partition(mtcars, c(test = 0.3, train = 0.7))
mod <- lm(mpg ~ wt, data = ex$train)
rmse(mod, ex$test)
rmse(mod, ex$train)

[Package modelr version 0.1.2 Index]