Generate an exclusive partitioning of a data frame

Generate an exclusive partitioning of a data frame

resample_partition(data, p)

Arguments

data
A data frame

p
A named numeric vector giving where the value is the probability that an observation will be assigned to that group.

Examples

ex <- resample_partition(mtcars, c(test = 0.3, train = 0.7))
mod <- lm(mpg ~ wt, data = ex$train)
rmse(mod, ex$test)

#> [1] 2.771096

rmse(mod, ex$train)

#> [1] 3.049972