Generate a sequence over the range of a vector

Generate a sequence over the range of a vector

Usage

seq_range(x, n, by, trim = NULL, expand = NULL, pretty = FALSE)

Arguments

x

A numeric vector

n, by

Specify the output sequence either by supplying the length of the sequence with n, or the spacing between value with by. Specifying both is an error.

I recommend that you name these arguments in order to make it clear to the reader.

trim

Optionally, trim values off the tails. trim / 2 * length(x) values are removed from each tail.

expand

Optionally, expand the range by expand * (1 + range(x) (computed after trimming).

pretty

If TRUE, will generate a pretty sequence. If n is supplied, this will use pretty() instead of seq(). If by is supplied, it will round the first value to a multiple of by.

Examples

x <- rcauchy(100)
seq_range(x, n = 10)
#>  [1] -84.84911 -61.76545 -38.68180 -15.59814   7.48552  30.56918  53.65284
#>  [8]  76.73649  99.82015 122.90381
seq_range(x, n = 10, trim = 0.1)
#>  [1] -11.5206327  -9.5557344  -7.5908360  -5.6259377  -3.6610393
#>  [6]  -1.6961410   0.2687574   2.2336557   4.1985541   6.1634524
seq_range(x, by = 1, trim = 0.1)
#>  [1] -11.5206327 -10.5206327  -9.5206327  -8.5206327  -7.5206327
#>  [6]  -6.5206327  -5.5206327  -4.5206327  -3.5206327  -2.5206327
#> [11]  -1.5206327  -0.5206327   0.4793673   1.4793673   2.4793673
#> [16]   3.4793673   4.4793673   5.4793673

# Make pretty sequences
y <- runif(100)
seq_range(y, n = 10)
#>  [1] 0.008551802 0.115207223 0.221862644 0.328518065 0.435173486
#>  [6] 0.541828907 0.648484328 0.755139749 0.861795171 0.968450592
seq_range(y, n = 10, pretty = TRUE)
#>  [1] 0.0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 1.0
seq_range(y, n = 10, expand = 0.5, pretty = TRUE)
#>  [1] -0.4 -0.2  0.0  0.2  0.4  0.6  0.8  1.0  1.2  1.4

seq_range(y, by = 0.1)
#>  [1] 0.008551802 0.108551802 0.208551802 0.308551802 0.408551802
#>  [6] 0.508551802 0.608551802 0.708551802 0.808551802 0.908551802
seq_range(y, by = 0.1, pretty = TRUE)
#>  [1] 0.0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 1.0